requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
stripe==6.6.0
python-dotenv==1.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Werkzeug==2.3.7
email-validator==2.0.0
redis==5.0.0
celery==5.3.4
